Fix type mismatch: use int for consecutive_command_failures_ to match max_consecutive_failures_
Agent-Logs-Url: https://github.com/chrominanzky/esphome-tesla-ble/sessions/b2e5e4b8-8b29-4133-b34d-dc4de8d27cc6

+++ components/tesla_ble_vehicle/tesla_ble_vehicle.h
@@ -268,7 +268,7 @@ namespace esphome
             int ble_disconnected_min_time_;
             int fast_poll_if_unlocked_ = 1; // != 0 enables fast polling
             int max_consecutive_failures_ = 3; // 0 = disabled, otherwise reboot after N consecutive command failures
-            uint8_t consecutive_command_failures_ = 0;
+            int consecutive_command_failures_ = 0;
             int number_updates_since_connection_ = 0;
             UniversalMessage_RoutableMessage read_queue_message_;
             CarServer_Response static_carserver_response_;
